requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pdfplumber>=0.10.0
langchain>=0.1.0
langchain-community>=0.0.10
ragas>=0.1.0
datasets>=2.14.0
pandas>=2.1.0
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0
aiofiles>=0.23.0 